requirements.txt
Appium-Python-Client==3.0.0
attrs==23.1.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
urllib3==2.0.5
wsproto==1.2.0
